+++ R/class_scheduler.R
@@ -22,12 +22,14 @@ scheduler_init <- function(
   )
   reporter <- reporter_init(reporter, seconds_interval = seconds_reporter)
   backoff <- tar_options$get_backoff()
+  trimmed <- counter_init()
   scheduler_new(
     graph = graph,
     queue = queue,
     progress = progress,
     reporter = reporter,
-    backoff = backoff
+    backoff = backoff,
+    trimmed <- trimmed
   )
 }
 
@@ -52,9 +54,10 @@ scheduler_new <- function(
   queue = NULL,
   progress = NULL,
   reporter = NULL,
-  backoff = NULL
+  backoff = NULL,
+  trimmed = NULL
 ) {
-  scheduler_class$new(graph, queue, progress, reporter, backoff)
+  scheduler_class$new(graph, queue, progress, reporter, backoff, trimmed)
 }
 
 scheduler_class <- R6::R6Class(
@@ -68,18 +71,21 @@ scheduler_class <- R6::R6Class(
     progress = NULL,
     reporter = NULL,
     backoff = NULL,
+    trimmed = NULL,
     initialize = function(
       graph = NULL,
       queue = NULL,
       progress = NULL,
       reporter = NULL,
-      backoff = NULL
+      backoff = NULL,
+      trimmed = NULL
     ) {
       self$graph <- graph
       self$queue <- queue
       self$progress <- progress
       self$reporter <- reporter
       self$backoff <- backoff
+      self$trimmed <- trimmed
     },
     count_unfinished_deps = function(name) {
       deps <- self$graph$produce_upstream(name)
@@ -98,12 +104,20 @@ scheduler_class <- R6::R6Class(
       self$progress$abridge()
       self$queue$abridge()
     },
+    trim = function(target, pipeline) {
+      parent_name <- target_get_parent(target)
+      parent_target <- pipeline_get_target(pipeline, parent_name)
+      downstream <- self$graph$produce_downstream(parent_name)
+      siblings <- target_get_children(parent_target)
+      counter_set_names(self$trimmed, c(downstream, siblings))
+    },
     validate = function() {
       self$graph$validate()
       self$queue$validate()
       self$progress$validate()
       self$reporter$validate()
       self$backoff$validate()
+      counter_validate(self$trimmed)
     }
   )
 )