+++ R/class_store_repository_cas.R
@@ -29,9 +29,13 @@ store_hash_late.tar_repository_cas <- function(store) {
 
 #' @export
 store_upload_object.tar_repository_cas <- function(store) {
+  store_upload_object_cas(store, store$file$stage)
+}
+
+store_upload_object_cas <- function(store, path) {
   on.exit(unlink(store$file$stage, recursive = TRUE, force = TRUE))
   tar_assert_scalar(
-    store$file$path,
+    path,
     msg = paste(
       "for a tar_repository_cas() target, the output must be",
       "a single file or single directory."
@@ -40,7 +44,7 @@ store_upload_object.tar_repository_cas <- function(store) {
   store_repository_cas_call_method(
     store = store,
     text = store$methods_repository$upload,
-    args = list(key = store$file$hash, path = store$file$stage)
+    args = list(key = store$file$hash, path = path)
   )
 }