+++ R/tar_target.R
@@ -195,13 +195,24 @@
 #'   stops and throws an error. Options:
 #'   * `"stop"`: the whole pipeline stops and throws an error.
 #'   * `"continue"`: the whole pipeline keeps going.
-#'   * `"abridge"`: any currently running targets keep running,
-#'     but no new targets launch after that.
-#'   (Visit <https://books.ropensci.org/targets/debugging.html>
-#'   to learn how to debug targets using saved workspaces.)
 #'   * `"null"`: The errored target continues and returns `NULL`.
 #'     The data hash is deliberately wrong so the target is not
 #'     up to date for the next run of the pipeline.
+#'   * `"abridge"`: any currently running targets keep running,
+#'     but no new targets launch after that.
+#'   * `"trim"`: all currently running targets stay running. A queued
+#'     target is allowed to start if:
+#'
+#'       1. It is not downstream of the error, and
+#'       2. It is not a sibling branch from the same [tar_target()] call
+#'         (if the error happened in a dynamic branch).
+#'
+#'     The idea is to avoid starting any new work that the immediate error
+#'     impacts. `error = "trim"` is just like `error = "abridge"`,
+#'     but it allows potentially healthy regions of the dependency graph
+#'     to begin running.
+#'   (Visit <https://books.ropensci.org/targets/debugging.html>
+#'   to learn how to debug targets using saved workspaces.)
 #' @param memory Character of length 1, memory strategy.
 #'   If `"persistent"`, the target stays in memory
 #'   until the end of the pipeline (unless `storage` is `"worker"`,